Change SasTransportImporter to trim trailing spaces from character values.
I had retained the strings because it was part of the XPORT file format, so it gave a truer view of the file that was parsed.  However, with feedback from developers, I decided to remove the trailing spaces since that's more natural in Java.   In short, this library is supposed to provide a convenient way to read and write SAS transport files from Java, not expose weird details about it.