Replace "data set" with "dataset".
This was done because "set" is also a transitive verb, which makes a name like "dataSetVariables" ambiguous: it could be setting the variables on the data, or it could be getting the data set's variables.  By combining "data" and "set" into a single word, it's more clearly recognized as a noun.

This follows the same convention used in the "sas7bdat" library, which will make reconciling their differences easier if the two libraries are ever combined.