Change "Anti-symmetry problems" from bold to a real section.
In this file's previous format, which was emacs org-file, it could have arbitrary nesting of sections.  In markdown it can't.  When converting from org to markdown, the end of the "anti-symmetry" section was lost.  By making it a real section, instead of a portion of the "design questions", it's clear where anti-symmetry problems end.