Interpret "null" as MissingValue.STANDARD when writing numeric values.
This change relaxes the "no nulls allowed" when exporting values in an observation to allow null for NUMERIC values.  The rationale is that other transports, like JSON or SQL databases, use null to indicate a missing numeric value, so it's unnecessarily burdensome to force a Java programmer to convert it to MissingValue.STANDARD when it's clear that that's the right thing to do.

A null value is still prohibited for CHARACTER values because all other transports support the empty string.

This makes the sas-transport library follow the pattern of the sas7bdat library, which may help reconcile their differences if they are ever combined into a single library.

The implementation is more invasive than the change suggests because it changes the exporter to be structured around the variable's type instead of the value's type.  This, again, follows the pattern in the sas7bdat library.

This also adds some more information to some IllegalArgumentException messages.