Allow writing LocalDate, LocalTime, LocalDateTime. Prohibit writing Date, Time, and Timestamp.
The problem with the java.sql.* time classes is that they have an implicit time zone, whereas SAS is given in "local time" (no time zone).  To convert a java.sql.* to local time, our code has to guess the time zone.  If it guesses wrong, it silently corrupts the data, which is much worse that forcing a programmer to convert to local time.

The implementation of this was adapted from the sas7bdat library.