Override MissingValue.toString()
This formats the MissingValue as a String in the was that SAS would format it.  Although this is not particularly useful, it does simply the sample code, since now all values and be printed as a String.

This also makes the MissingValue behave the same as the MissingValue class in the sas7bdat library, which may help if the two libraries are ever combined into one.