Reduce the version to 0.9.0.
After thinking about it, I don't think it's right to claim that this is a 1.0 release.  The library has only been used once, and only for writing character ad numeric data (no dates).  It has never been used for reading.  Also, the API is a bit rough, since it has complex constructors instead of builders.